Move vscode-common's TestHelpers interface to its own file.
